@@ -72,14 +72,14 @@ namespace DotPulsar
         /// </summary>
         public int BatchIndex => Data.BatchIndex;
 
-        public override bool Equals(object o)
+        public override bool Equals(object? o)
             => o is MessageId id && Equals(id);
 
-        public bool Equals(MessageId other)
-            => !(other is null) && LedgerId == other.LedgerId && EntryId == other.EntryId && Partition == other.Partition && BatchIndex == other.BatchIndex;
+        public bool Equals(MessageId? other)
+            => other is not null && LedgerId == other.LedgerId && EntryId == other.EntryId && Partition == other.Partition && BatchIndex == other.BatchIndex;
 
         public static bool operator ==(MessageId x, MessageId y)
-            => ReferenceEquals(x, y) || (x is object && x.Equals(y));
+            => ReferenceEquals(x, y) || (x is not null && x.Equals(y));
 
         public static bool operator !=(MessageId x, MessageId y)
             => !(x == y);
